middle cahnges

+++ dev/LogicLayer/OrderLine.java
@@ -5,11 +5,11 @@ public class OrderLine {
     private int itemId;
     private String itemName;
     private int itemQuantity;
-    private int itemCost;
-    private int itemDiscount ;
-    private int finalCost;
+    private Double itemCost;
+    private Double itemDiscount ;
+    private Double finalCost;
 
-    public OrderLine(int itemId, String itemName ,int itemQuantity , int  itemCost, int itemDiscount , int finalCost )
+    public OrderLine(int itemId, String itemName ,int itemQuantity , Double  itemCost, Double itemDiscount , Double finalCost )
     {
         this.itemId=itemId;
         this.itemName= itemName;
